+++ frontend/src/core/registry.js
@@ -24,6 +24,8 @@ const MODULES = {
   'skew-up':     () => import('../modules/skew-up.js'),
   'hover-zoom':   () => import('../modules/hover-zoom.js'),
   'img-parallax': () => import('../modules/img-parallax.js'),
+  'page-fade':    () => import('../modules/page-fade.js'),
+  'page-curtain': () => import('../modules/page-curtain.js'),
 };
 
 export async function loadModules(activeModules) {